Modifying load_iaga_2000() parsing algorithm.
Algorithm modified to use whitepace to split data columns by default,
falling back to using the position of the column headers if the number
of columns isn't as expected. This approach is taken to fix a parsing
problem observed with the University of Tromso magnetometer data from
Tristan da Cunha. The position-based approach used previously is kept
as a fallback in case of circumstances where the columns expand to
full width leaving no whitespace.