Add generic save function
Save data. If path not given then save to default file(s) as defined by
the data type and archive. If data would span multiple files it is split
and each part saved separately. By default data saved to the default
location is merged with existing data; overriding any samples which
already exist or adding values for new timestamps.

When saving to a specified path the save_converter must be specified. It
is possible to merge data with existing samples by specifying
merge=True.