Add/use generic load converter
Generalised load converter to read in plain text data files.
Requires some additional information to be added into the archive
data structure to define the method used for timestamping the
data (eg offsets from start time or unix timestamps). It also
requires a reference to the class constructor in order to return
the correct data type. When units are T it is assumed that the
data files are written with values in nT for readability.