Add function to generate path for realtime MagQDC data
For the realtime case a previous monthly QDC must be used. For the
first few days of the month assume the previous month's QDC does not
exist and use the one prior to that. Actual path to use is defined
is a different archive (for the 'realtime_qdc' archive the path in
from the 'qdc' archive is used.