Add realtime option to load_qdc to adjust file used
When realtime mode is enabled it is assumed that the QDC for the
date specified is (was) not available and that the previous month's
QDC should be used. For days in the month before qdc_rollover it is
also assumed that the previous month's QDC is (was) not available
and to use the QDC from the month before that.

With the realtime option the realtime_qdc archives are no longer
required.