Enable cadence to be set after data loaded
The --cadence option sets the cadence as each data file is loaded.
Adjusting this value alters how filtered/processed-on-demand data
is treated. Use  --cadence when loading long time intervals where
is is important to minimise peak memory usage.

--post-cadence sets the cadence after all of the data has been
loaded, it is does not affect how filtered/processed-on-demand
data is treated.